+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Rename and Backlinks no longer use id and aliases to identify notes, to stay consistent with obsidian app.
+
+### Fixed
+
+- Rename will not error on different id and file basename.
+
 ## [v3.14.3](https://github.com/obsidian-nvim/obsidian.nvim/releases/tag/v3.14.3) - 2025-10-19
 
 ### Fixed